pyproject.toml
[project]
name = "digital-workforce"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.92.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "supabase>=2.10.0",
    "apify-client>=1.8.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "reportlab>=4.0.0",
    "arq>=0.25.0",
    "redis>=5.0.0",
    "structlog>=24.0.0",
    "playwright>=1.49.0",
    "openai>=2.0.0",
    "browserbase>=1.0.0",
    "selenium>=4.0.0",
    "firecrawl-py>=1.0.0",
]

[project.optional-dependencies]
dev = ["ruff", "mypy", "pytest", "pytest-asyncio", "httpx"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
